Feature/booking flight

+++ src/main/java/com/dvk/ct250backend/domain/flight/dto/FlightOverview.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -13,6 +15,6 @@ import lombok.experimental.FieldDefaults;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class FlightOverview {
     String date;
-    double minPriceOfDay;
+    BigDecimal minPriceOfDay;
     boolean hasFlight;
 }
